Fix the build

+++ app/src/main/kotlin/android/emu6502/nes/mappers/AOROM.kt
@@ -1,13 +1,13 @@
 package android.emu6502.nes.mappers
 
 class AOROM : Mapper {
-  override fun write(address: Int, value: Byte) {
+  override fun write(address: Int, value: Int) {
   }
 
   override fun step() {
   }
 
-  override fun read(address: Int): Byte {
+  override fun read(address: Int): Int {
     throw NotImplementedError()
   }
 }
